+++ src/fragments/map-view/MapView.vue
@@ -50,6 +50,7 @@
             <l-popup v-if="polygon.properties.label">
               <div :ref="'isochronePopupContainer' + index" >
                 {{polygon.properties.label}} {{$t('mapView.polygon')}} - {{polygon.properties.area}} 
+                <br/><span>{{$t('global.population')}}: {{polygon.properties.total_pop}}</span>
               </div>
             </l-popup>
           </l-polygon>